backend/requirements.txt
fastapi
python-dotenv
requests
sqlalchemy
uvicorn
aio-pika
httpx
openai
Pillow
scikit-image
imagehash
numpy
opencv-python-headless
pywebpush
pydantic-settings
pytesseract
aiohttp
rapidfuzz
python-multipart